requirements.txt
redis==4.5.5
elasticsearch[async]==8.7.0
fastapi==0.95.2
orjson==3.8.13
pydantic==1.10.8
uvicorn==0.22.0
uvloop==0.17.0 ; sys_platform != "win32" and implementation_name == "cpython"
flake8==6.0.0
pytest==7.3.1
python-dotenv==1.0.0
gunicorn==20.1.0
